+++ src/main/java/roidrole/thaumicinfo/ThaumicInformationConfig.java
@@ -6,6 +6,7 @@ import net.minecraftforge.common.config.ConfigManager;
 import net.minecraftforge.fml.client.event.ConfigChangedEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+import roidrole.thaumicinfo.jei.ResearchManager;
 
 @Config(
 	modid = Tags.MOD_ID,
@@ -72,7 +73,6 @@ public class ThaumicInformationConfig {
 	@Config.Name("JEI Configs")
 	public static final JEI jeiConfig = new JEI();
 	public static class JEI {
-		@Config.RequiresMcRestart
 		@Config.Comment("Hide recipes from JEI if you don't have the research for it")
 		public boolean hideRecipesIfMissingResearch = false;
 
@@ -197,6 +197,11 @@ public class ThaumicInformationConfig {
 		public static void onConfigChanged(ConfigChangedEvent.OnConfigChangedEvent event) {
 			if (event.getModID().equals(Tags.MOD_ID)) {
 				ConfigManager.sync(Tags.MOD_ID, Config.Type.INSTANCE);
+				if(ThaumicInformationConfig.jeiConfig.hideRecipesIfMissingResearch){
+					ResearchManager.init();
+				} else {
+					ResearchManager.deInit();
+				}
 			}
 		}
 	}